Add all interface listen for rack-app
In order to ensure we can properly route requests within a kubernetes
cluster to our example rack-app, this commit changes the Dockerfile to
listen on all interfaces (0.0.0.0) instead of localhost. This is
necessary as within a kubernetes cluster, the request won't originate
within the pod itself which means we won't be communicating with our pod
properly.

Additionally, we hardcode the project name in the docker-build-image
script as the other files expect the project name to be rack-app. This
should handle the case where the project directory name can't break the
expectations of the example project.